+++ web/src/actions/app/navigation.js
@@ -1,28 +1,8 @@
 import { push } from 'connected-react-router';
-import { NAVIGATION_SCROLL_TO_WAYPOINT, RESET } from './symbols';
+import { RESET } from './symbols';
 
-let jumping = false;
-
+// eslint-disable-next-line import/prefer-default-export
 export const goToDashboard = ({ pushRoute = push } = {}) => dispatch => {
   dispatch({ type: RESET });
   dispatch(pushRoute('/'));
-};
-
-export const jumpTo = waypoint => dispatch => {
-  if (waypoint) {
-    jumping = true;
-    setTimeout(() => {
-      jumping = false;
-    }, 100);
-    dispatch({ type: NAVIGATION_SCROLL_TO_WAYPOINT, data: waypoint });
-  }
-};
-
-export const scrollTo = waypoint => dispatch => {
-  if (!jumping && waypoint) {
-    dispatch({
-      type: NAVIGATION_SCROLL_TO_WAYPOINT,
-      data: waypoint
-    });
-  }
 };