+++ CHANGELOG.md
@@ -5,8 +5,16 @@ Anticipated release: August 10, 2020
 #### 🚀 New features
 
 - Upgrade to @cmsgov/design-system v2.0.0, remove focus styles ([#2412])
+- Update Key Personnel and Program Management page instructions and help text ([#2325])
 
 #### 🐛 Bugs fixed
+
+- Key personnel FTE not being saved for 2022 ([#2401])
+- Arrows in the side nav should be consistent ([#2188])
+- Nav bar does not indicate correct location with launch of different APD ([#2223])
+- Clicking on the navigation places you right below the anchor point ([#2410])
+- Scroll changes navigation buttons on export and submit page ([#2359])
+- When navigating with Continue/Previous buttons, Nav does not expand to current page ([#2415])
 
 #### ⚙️ Behind the scenes
 
@@ -16,5 +24,12 @@ Anticipated release: August 10, 2020
 
 See our [release history](https://github.com/18F/cms-hitech-apd/releases)
 
-[#2335]: https://github.com/18F/cms-hitech-apd/issues/2335
 [#2412]: https://github.com/CMSgov/eAPD/issues/2412
+[#2335]: https://github.com/CMSgov/eAPD/issues/2335
+[#2325]: https://github.com/CMSgov/eAPD/issues/2325
+[#2401]: https://github.com/CMSgov/eAPD/issues/2401
+[#2188]: https://github.com/CMSgov/eAPD/issues/2188
+[#2223]: https://github.com/CMSgov/eAPD/issues/2223
+[#2410]: https://github.com/CMSgov/eAPD/issues/2410
+[#2359]: https://github.com/CMSgov/eAPD/issues/2359
+[#2415]: https://github.com/CMSgov/eAPD/issues/2415